An overhaul of how IOManagers keep track of open IOStreams and their corresponding IOMethods.
Automatically truncate files when writing if no offset was specified and no previous I/O has occurred.
This is how HyperTalk does it, this is (now) the proper behavior as specified by the Standard (previously it was left undefined),
this is more intuitive, and the fact that previous versions did not do this really bugged me.
Compatibility with previous versions is NOT an excuse, especially at this point where I only have two users, myself included.
Case in point: The make utility will not work correctly unless command lines are preceded by a single tab.
Spaces and mixed spaces and tabs will not cut it. This has caused thousands of people headaches over the years.
It could have been fixed a long time ago in the first version, but the author didn't want to because he already had ten users.

Fixes Issue 9.